// api/src/domain/ContainerState.ts
// domain/ContainerState.ts
import { BaseNode } from "./BaseNode";

export interface Composition {
  [component: string]: number; // e.g. { cabernet: 0.7, merlot: 0.3 }
}

export interface ContainerState extends BaseNode {
  containerId: string;
  operationId: string;
  previousStateId?: string;
  volumeLiters: number;
  composition: Composition;
  timestamp: Date;
  isCurrent: boolean;
  isInitial?: boolean;
}